update conditionals, spelling error

+++ setup.sh
@@ -19,7 +19,7 @@ fi
 # Moving Xdefaults into place
 read -p "Install Xdefaults? (y/n): " response
 if [[ "$response" == "y" ]]; then
-	cp ~/.Xdefaults ~/.Xdefaults.bak
+	if [ -f ~/.Xdefaults ]; then cp ~/.Xdefaults ~/.Xdefaults.bak; fi
 	cp -f ./Xdefaults ~/.Xdefaults
 	echo "setup complete:	.Xdefaults"
 elif [[ "$response" == "n" ]]; then
@@ -31,7 +31,7 @@ fi
 # Moving tmux.conf into place
 read -p "Install tmux.conf? (y/n): " response
 if [[ "$response" == "y" ]]; then
-	cp ~/.tmux.conf ~/.tmux.conf.bak
+	if [ -f ~/.tmux.conf ]; then cp ~/.tmux.conf ~/.tmux.conf.bak; fi
 	cp -f ./tmux.conf ~/.tmux.conf
 	echo "setup complete:	.tmux.conf"
 elif [[ "$response" == "n" ]]; then
@@ -43,7 +43,7 @@ fi
 # Moving spectrwm.conf into place
 read -p "Install spectrwm.conf? (y/n): " response
 if [[ "$response" == "y" ]]; then
-	cp ~/.spectrwm.conf ~/.spectrwm.bak
+	if [ -f ~/.tmux.conf ]; then cp ~/.spectrwm.conf ~/.spectrwm.bak; fi
 	cp -f ./spectrwm.conf ~/.spectrwm.conf
 	echo "setup complete:	.spectrwm.conf"
 elif [[ "$response" == "n" ]]; then
@@ -55,7 +55,7 @@ fi
 # Moving i3wm config into place
 read -p "Install i3 config? (y/n): " response
 if [[ "$response" == "y" ]]; then
-	cp -r ~/.config/i3 ~/.config/i3.bak
+	if [ -f ~/.config/i3 ]; then cp -r ~/.config/i3 ~/.config/i3.bak; fi
 	cp -rf i3 ~/.config/.
 	echo "setup complete:	i3 config"
 elif [[ "$response" == "n" ]]; then
@@ -67,7 +67,7 @@ fi
 # Moving fluxbox config into place
 read -p "Install fluxbox config? (y/n): " response
 if [[ "$response" == "y" ]]; then
-	cp -r ~/.config/fluxbox ~/.config/fluxbox.bak
+	if [ -f ~/.config/fluxbox ]; then cp -r ~/.config/fluxbox ~/.config/fluxbox.bak; fi
 	cp -rf fluxbox ~/.config/.
 	echo "setup complete:	fluxbox config"
 elif [[ "$response" == "n" ]]; then
@@ -79,7 +79,7 @@ fi
 # Moving vim config into place
 read -p "Install vim config? (y/n): " response
 if [[ "$response" == "y" ]]; then
-	cp -r ~/.vimrc ~/.vimrc.bak
+	if [ -f ~/.vimrc ]; then cp -r ~/.vimrc ~/.vimrc.bak; fi
 	cp -rf vimrc ~/.vimrc
 	echo "setup complete:	vim config"
 elif [[ "$response" == "n" ]]; then
